Allow SQL Server identity columns for byte properties, but warn
Issue #7161

Byte properties are set to not be store-generated by convention, but if store-generation is configured, then SQL Server will use an identity column and a warning is issued in model validation.

Also a small fix to the warning for decimal properties to ensure that nullable decimals also get the warning if they have no explicit mapping.